refactor: invert CLI entry and lazy-load the MCP transport out of the eager graph
